Move save effects to descendant Settings models
