+++ apps/core/models.py
@@ -1,6 +1,5 @@
 from typing import Any, Union
 
-from django.apps import apps
 from django.contrib import admin
 from django.core.exceptions import ValidationError
 from django.db import models
@@ -222,34 +221,6 @@ class Setting(BaseModel):
         SettingFieldType.EMAIL: "email",
     }
 
-    TRIGGERS = {
-        "festival_status": {
-            True: {
-                "SettingMain": {
-                    "main_add_blog": True,
-                    "main_add_news": False,
-                    "main_add_places": True,
-                    "main_add_short_list": False,
-                    "main_show_afisha_only_for_today": True,
-                }
-            },
-            False: {
-                "SettingMain": {
-                    "main_add_blog": False,
-                    "main_add_news": True,
-                    "main_add_places": False,
-                    "main_add_short_list": True,
-                    "main_show_afisha_only_for_today": False,
-                }
-            },
-        }
-    }
-    do_not_check_triggers = False
-
-    RELATED_SETTINGS = {
-        "main_add_blog": "main_add_news",
-        "main_add_news": "main_add_blog",
-    }
     HELP_TEXT = {
         "main_add_blog": BLOG_HELP_TEXT,
         "main_add_news": NEWS_HELP_TEXT,
@@ -317,12 +288,6 @@ class Setting(BaseModel):
                 {"image": "Изображение должно присутствовать на странице. Оставьте или замените на другое."}
             )
 
-    def save(self, *args, **kwargs):
-        self._check_related_settings(self)
-        if not type(self).do_not_check_triggers:
-            self._check_triggers(self)
-        super().save(*args, **kwargs)
-
     @property
     def value(self):
         return getattr(
@@ -359,26 +324,7 @@ class Setting(BaseModel):
 
         return settings_dict
 
-    @classmethod
-    def _set_setting(cls, setting, value=False):
-        if cls.objects.filter(settings_key=setting).exists():
-            setting = cls.objects.get(settings_key=setting)
-            setting.boolean = value
-            setting.save()
-
-    @classmethod
-    def _check_related_settings(cls, setting):
-        if setting.settings_key in cls.RELATED_SETTINGS and setting.boolean:
-            cls._set_setting(cls.RELATED_SETTINGS[setting.settings_key])
-
-    @classmethod
-    def _check_triggers(cls, setting):
-        if setting.settings_key in cls.TRIGGERS:
-            preset = cls.TRIGGERS[setting.settings_key]
-            for model, settings in preset.get(setting.boolean, {}).items():
-                for key, value in settings.items():
-                    cls.do_not_check_triggers = True
-                    try:
-                        apps.get_model("main", model)._set_setting(key, value)
-                    finally:
-                        cls.do_not_check_triggers = False
+    def _set_settings(self, settings):
+        for key, value in settings.items():
+            count = Setting.objects.filter(settings_key=key).update(boolean=value)
+            assert count == 1, f"Количество записей с ключом '{key}' оказалось равно {count} (ожидалось 1)"
